[Av][47][Done]

+++ resources/views/kurban/show.blade.php
@@ -77,7 +77,8 @@
 
                     @if ($kurban->kurbanPeserta->count() == 0)
                         <div class="text-center">Belum ada data.
-                            <a href="{{ route('kurbanpeserta.create', ['kurban_id' => $kurban->id]) }}">Pendaftaran Baru</a>
+                            <a href="{{ route('kurbanpeserta.create', ['kurban_id' => $kurban->id]) }}">Pendaftaran
+                                Baru</a>
                         </div>
                     @else
                         <table class="table table-striped">
@@ -109,24 +110,32 @@
                                         </td>
                                         <td>
                                             @if ($item->status_bayar == 'lunas')
-                                                <span class="badge bg-success me-1 my-1">LUNAS</span>
+                                                <span
+                                                    class="badge bg-success me-1 my-1">{{ $item->getStatusText() }}</span>
                                             @else
-                                                <span class="badge bg-secondary me-1 my-1">BELUM LUNAS</span>
+                                                <span
+                                                    class="badge bg-secondary me-1 my-1">{{ $item->getStatusText() }}</span>
                                             @endif
                                         </td>
                                         <td>
-                                            {!! Form::open([
-                                                'method' => 'DELETE',
-                                                'route' => ['kurbanpeserta.destroy', [$item->id, 'kurban_id' => $item->kurban_id]],
-                                                'style' => 'display:inline',
-                                            ]) !!}
-                                            @csrf
-                                            @method('DELETE')
-                                            <a href="{{ route('kurbanpeserta.edit', [$item->id, 'kurban_id' => $item->kurban_id]) }}"
-                                                class="btn btn-primary btn-sm mb-1 mx-1">Pembayaran Iuran</a>
-                                            <button type="submit" class="btn btn-danger btn-sm mb-1 mx-1"
-                                                onclick="return confirm('Apakah Anda yakin ingin menghapus data ini?')">Hapus</button>
-                                            {!! Form::close() !!}
+                                            @if ($item->status_bayar != 'lunas')
+                                                {!! Form::open([
+                                                    'method' => 'DELETE',
+                                                    'route' => ['kurbanpeserta.destroy', [$item->id, 'kurban_id' => $item->kurban_id]],
+                                                    'style' => 'display:inline',
+                                                ]) !!}
+                                                @csrf
+                                                @method('DELETE')
+                                                <a href="{{ route('kurbanpeserta.edit', [$item->id, 'kurban_id' => $item->kurban_id]) }}"
+                                                    class="btn btn-primary btn-sm mb-1 mx-1">Pembayaran Iuran</a>
+                                                <button type="submit" class="btn btn-danger btn-sm mb-1 mx-1"
+                                                    onclick="return confirm('Apakah Anda yakin ingin menghapus data ini?')">Hapus</button>
+                                                {!! Form::close() !!}
+                                            @else
+                                                Sudah Lunas
+                                            @endif
+
+
                                         </td>
                                     </tr>
                                 @endforeach
